requirements.txt
asgiref==3.3.4
Django==3.2
django-admin-material==0.1.0
django-crispy-forms==1.11.2
django-materializecss-form==1.1.17
djangorestframework==3.12.4
gunicorn==20.1.0
pytz==2021.1
sqlparse==0.4.1
whitenoise==5.2.0
